Spread django test units into multiple files. For further readabilities.
